+++ examples/simple_example.cc
@@ -1,57 +1,20 @@
 // Copyright (c) 2020 Aleksey Timin <atimin@gmail.com>
 // Licensed under the MIT License.
 
-#include <ntio/sockets/tcp_socket.h>
+#include <ntio/core/error.h>
+#include <ntio/core/logger.h>
+#include <ntio/core/task.h>
 #include <ntio/sockets/tcp_listener.h>
+#include <ntio/sockets/tcp_socket.h>
 
 #include <coroutine>
 #include <iostream>
 #include <memory>
 
 static const int kPort = 6545;
-struct Promise;
 
-struct Task {
-  using promise_type = Promise;
-
-  explicit Task(int& state) : state_(state) {}
-  int& state_;
-
-  operator int() { return state_; }
-};
-
-struct Promise {
-  Task get_return_object() {
-    std::cout << "Promise.get_return_object()" << std::endl;
-    return Task(state_);
-  }
-
-  std::suspend_never initial_suspend() {
-    std::cout << "Promise.initial_suspend()" << std::endl;
-    return {};
-  }
-
-  std::suspend_never final_suspend() noexcept {
-    std::cout << "Promise.final_suspend()" << std::endl;
-    return {};
-  }
-
-  void return_void() { std::cout << "Promise.return_void()" << std::endl; }
-
-  void return_value(const int& val) {
-    std::cout << "Promise.return_value() with " << val << std::endl;
-    state_ = val;
-  }
-
-  std::suspend_never yield_value(const int& val) {
-    std::cout << "Promise.yield_value() with " << val << std::endl;
-    return {};
-  }
-
-  void unhandled_exception() { std::cout << "Promise.unhandled_exception()" << std::endl; }
-
-  int state_;
-};
+using ntio::core::Error;
+using Task = ntio::core::Task<Error>;
 
 Task PingPong() {
   using ntio::sockets::EndPoint;
@@ -60,8 +23,8 @@ Task PingPong() {
 
   TcpListener listener;
   if (auto err = co_await listener.Bind(EndPoint("127.0.0.1", kPort))) {
-    std::cout << "Bind error: " << err << std::endl;
-    co_return;
+    LOG_ERROR("Bind error: " << err);
+    co_return err;
   }
 
   auto accept_task = listener.Accept();
@@ -70,14 +33,14 @@ Task PingPong() {
   auto con_task = client.Connect(EndPoint("127.0.0.1", kPort));
 
   if (auto err = co_await con_task) {
-    std::cout << "Conn error: " << err << std::endl;
-    co_return;
+    LOG_ERROR("Conn error: " << err);
+    co_return err;
   }
 
   auto [sock, accept_err] = co_await accept_task;
   if (accept_err) {
-    std::cout << "Accept error: " << accept_err << std::endl;
-    co_return;
+    LOG_ERROR("Accept error: " << accept_err);
+    co_return accept_err;
   }
 
   std::vector<uint8_t> rbuffer(1'000'000);
@@ -85,22 +48,24 @@ Task PingPong() {
 
   auto [write_bytes, write_err] = co_await sock->Write(wbuffer, 1'000'000);
   if (write_err) {
-    std::cout << "Write error: " << write_err << std::endl;
-    co_return;
+    LOG_ERROR("Write error: " << write_err);
+    co_return write_err;
   }
 
-  std::cout << "Write " << write_bytes << " bytes" << std::endl;
+  LOG_INFO("Write " << write_bytes << " bytes");
 
   auto [read_bytes, read_err] = co_await client.Read(&rbuffer, 1'000'000);
   if (read_err) {
-    std::cout << "Read error: " << read_err << std::endl;
-    co_return;
+    LOG_ERROR("Read error: " << read_err);
+    co_return read_err;
   }
 
-  std::cout << "Read " << read_bytes << " bytes" << std::endl;
+  LOG_INFO("Read " << read_bytes << " bytes");
+  co_return Error();
 }
 
 int main() {
-  PingPong();
-  return 0;
+  auto ret = PingPong();
+  LOG_INFO("Result: " << ret);
+  return ret ? -1 : 0;
 }