+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 * Add documentation, #1 by @flipback
 * Add GitHub acctions to build and lint code, #2 by @flipback
 * Fix log macros, #3 by @flipback
-* Move Logger and Errro to ntio::core, #4 by @flipback
+* Move Logger and Error to ntio::core, #4 by @flipback
+* Add Task class, #5 by @flipback
 
 # Release 0.1.0 (29-12-2020)